project is moduled. utils.js and data.js added
